requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
black==25.9.0
certifi==2025.8.3
click==8.3.0
frozenlist==1.7.0
idna==3.10
magic-filter==1.0.12
multidict==6.6.4
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
propcache==0.3.2
pydantic==2.11.10
pydantic_core==2.33.2
python-dotenv==1.1.1
pytokens==0.1.10
tomli==2.2.1
typing-inspection==0.4.2
typing_extensions==4.15.0
yarl==1.20.1
